Refactor user registration enums by moving UserRegisterResult and Role to enums package
