+++ src/main/java/oneseoktwojo/ohtalkhae/domain/auth/enums/Role.java
@@ -1,8 +1,8 @@
-package oneseoktwojo.ohtalkhae.domain.auth;
+package oneseoktwojo.ohtalkhae.domain.auth.enums;
 
 public enum Role {
     ROLE_USER, ROLE_ADMIN;
-    
+
     @Override
     public String toString() {
         return name();
